+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Fixed
 
 -  Fix `Get Conda Eenvs Error: /bin/sh: 1: conda: not found`
+-  Fix conda env select
 
 ### Added